Update TemplateData and VisualEditor submodules: Add "suggestedvalues" property
Parameters now contain a "suggestedvalues" property. A list of suggested values
can be set when managing template data and they can be selected through a
dropdown menu when calling the template in VisualEditor.

The following variables must be set in LocalSettings.php to make it work:
$wgTemplateDataSuggestedValuesEditor = true;
$wgVisualEditorTransclusionDialogSuggestedValues = true;

Extensions TemplateData and VisualEditor are thus updated